Rename  helper to entity_helper
